Extract hook for add-to-wishlist
